more compatibility to windows versions, broke something in multithreading...

+++ src/headers/numberHelper.h
@@ -8,7 +8,7 @@
 #include "getDataObjects.h"
 
 size_t roundUp8(size_t numToRound);
-uint64_t getBytesAsNumber(byte* chunk_start, size_t num_bytes, ByteOrder endianness);
+size_t getBytesAsNumber(byte* chunk_start, size_t num_bytes, ByteOrder endianness);
 void reverseBytes(byte* data_pointer, size_t num_elems);
 
 #endif //NUMBER_HELPER_H
